requirements.txt
semantic-kernel
fastapi
uvicorn
asyncpg
sqlalchemy
pandas
tabulate
python-dotenv
pydantic